% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tracking.R
\name{affected_downstream_basins}
\alias{affected_downstream_basins}
\title{Select downstream basins affected by water-right changes}
\usage{
affected_downstream_basins(changed_pods, basins, comid = "comid")
}
\arguments{
\item{changed_pods}{An \code{sf} object returned by \code{\link[=detect_water_right_changes]{detect_water_right_changes()}}.}

\item{basins}{An \code{sf} object of upstream COMID basins from \code{\link[=get_pod_basin]{get_pod_basin()}}.}

\item{comid}{Column in \code{basins} that contains COMIDs.}
}
\value{
An \code{sf} subset of \code{basins} that contains the changed PODs.
}
\description{
Select downstream basins affected by water-right changes
}
