% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pod_basin}
\alias{get_pod_basin}
\title{Get POD basins}
\usage{
get_pod_basin(comid, crs)
}
\arguments{
\item{crs}{A \code{sf::st_crs()} object.}

\item{data}{A previously filter flowmet object}
}
\description{
Get POD basins
}
