% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepped_water_rights.R
\name{read_prepped_water_rights}
\alias{read_prepped_water_rights}
\title{Read a prepped water-right GeoPackage cache}
\usage{
read_prepped_water_rights(path, layer = "water_rights_prepped")
}
\arguments{
\item{path}{GeoPackage created by \code{\link[=write_prepped_water_rights]{write_prepped_water_rights()}}.}

\item{layer}{Layer name to read.}
}
\value{
Prepared canonical water-right records.
}
\description{
Read a prepped water-right GeoPackage cache
}
