% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tracking.R
\name{read_water_right_snapshot}
\alias{read_water_right_snapshot}
\title{Read a water-right snapshot}
\usage{
read_water_right_snapshot(path)
}
\arguments{
\item{path}{A GeoPackage created by \code{\link[=write_water_right_snapshot]{write_water_right_snapshot()}}.}
}
\value{
An \code{sf} object.
}
\description{
Read a water-right snapshot
}
