% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_right_schema.R
\name{standardize_mt_water_rights}
\alias{standardize_mt_water_rights}
\title{Standardize Montana water rights}
\usage{
standardize_mt_water_rights(pods)
}
\arguments{
\item{pods}{An \code{sf} object from Montana's \code{WRQS_PODS} layer.}
}
\value{
Canonical WateRAT water-right records.
}
\description{
Standardize Montana water rights
}
