% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_right_schema.R
\name{validate_water_rights}
\alias{validate_water_rights}
\title{Validate canonical water-right records}
\usage{
validate_water_rights(water_rights)
}
\arguments{
\item{water_rights}{An \code{sf} object to validate.}
}
\value{
\code{water_rights}, invisibly.
}
\description{
Validate canonical water-right records
}
