% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idaho_water_rights.R
\name{write_idwr_pods_by_use}
\alias{write_idwr_pods_by_use}
\title{Write expanded Idaho water uses}
\usage{
write_idwr_pods_by_use(pods_by_use, path, layer = "PODRightWaterUses")
}
\arguments{
\item{pods_by_use}{An \code{sf} object from \code{\link[=expand_idwr_pods_by_use]{expand_idwr_pods_by_use()}}.}

\item{path}{Output GeoPackage or writable File Geodatabase path.}

\item{layer}{Output layer name.}
}
\value{
\code{path}, invisibly.
}
\description{
Write expanded Idaho water uses
}
